backend/requirements.txt
fastapi>=0.104.1,<0.105
uvicorn[standard]>=0.23.2,<0.24
spacy>=3.7.2,<3.8
scikit-learn>=1.3.2,<1.4
requests>=2.31.0,<2.32
python-dotenv>=1.0.0,<1.1
GitPython>=3.1.40,<3.2 